broken link on wifi.md
The link to the WifiInterface API documentation is broken at the top of the page.  You can find an API doc for WifiInterface through the search box at the following location: https://docs.mbed.com/docs/vignesh/en/latest/api/classWiFiInterface.html

but I am not sure if this is the correct doc.